resume-screener-filled/requirements.txt
# Streamlit app
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.1
qrcode==7.4.2

# NLP
sentence-transformers==2.2.2
torch==2.2.2

# File parsing
python-docx==1.1.0
pdfminer.six==20221105

# PDF export
reportlab==4.2.2